refactor: fetchTags axios로 변경 후 적용
